+++ src/services/import/index.ts
@@ -1,3 +1,4 @@
 export * from "@/services/import/import.service";
 export * from "@/services/import/extension-points";
 export * from "@/services/import/dto";
+export * from "@/services/import/profiling";